### 2017-08-01 #344 @whedon
[PRE REVIEW]: The psyplot interactive visualization framework
**Submitting author:** @Chilipp (<a href="http://orcid.org/0000-0001-6171-7716">Philipp S. Sommer</a>)
**Repository:** <a href="https://github.com/Chilipp/psyplot.git" target ="_blank">https://github.com/Chilipp/psyplot.git</a>
**Version:** v1.0.0
