### 2017-08-02 309c5ed
Added table for download of latest psyplot-conda binaries
