### 2017-08-03 1e43c8a
Added note about command line download [skip ci]
