### 2017-08-03 9925f29
Minor changes [skip ci]
