### 2017-08-03 714d27d
Added instructions about installer usage [skip ci]
