### 2017-08-04 ea83452
Use static links to the psyplot-conda binaries [skip ci]
