### 2017-08-05 a7ebd9e
Minor changes in classifiers [skip ci]
