### 2017-08-05 0739b1d
Use conda-forge repositories [skip ci]
