### 2017-08-12 644c76f
Do only install from conda-forge [skip ci]
