### 2017-08-12 607bc8f
Improve installation instructions [skip ci]
