+++ docs/installing.rst
@@ -8,8 +8,10 @@ Installation
 How to install
 --------------
 
-Installation using single executable
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+.. _install-standalone:
+
+Installation via standalone installers
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
 .. only:: html and not epub
 
@@ -17,8 +19,8 @@ Installation using single executable
     | | |win-logo| | | |osx-logo| | | |linux-logo| |
     | | Windows    | | Mac OS X   | | Linux        |
     +==============+==============+================+
-    | | |win-64|   | |osx-64|     | |linux-64|     |
-    | | |win-32|   |              |                |
+    | | |win-64|   | | |osx-pkg|  | |linux-64|     |
+    | | |win-32|   | | |osx-64|   |                |
     +--------------+--------------+----------------+
 
     .. |win-logo| image:: windows.png
@@ -31,15 +33,25 @@ Installation using single executable
 
     .. |win-32| replace:: :psycon:`32-bit (.exe installer) <Windows-x86.exe>`
 
+    .. |osx-pkg| replace:: :psycon:`64-bit (.pkg installer) <MacOSX-x86_64.pkg>`
+
     .. |osx-64| replace:: :psycon:`64-bit (bash installer) <MacOSX-x86_64.sh>`
 
     .. |linux-64| replace:: :psycon:`64-bit (bash installer) <Linux-x86_64.sh>`
 
-The easiest way for the installation comes via the psyplot-conda_ project.
-Here we provide ready-to-go installers for psyplot for all
-platforms. Those installers include conda, psyplot, the
-`graphical user interface`_, several plugins (psy-simple_,
-psy-maps_ and psy-reg_) and their dependencies.
+
+This section contains the necessary informations to install psyplot-conda, a
+standalone psyplot installation with the most important plugins and the
+graphical user interface (GUI).
+
+Executables can be downloaded from the links above. Older versions are also
+available through the the releases_ page, nightly builds for Linux and OSX are
+available here_.
+
+The installer provided here contain all necessary dependencies for psyplot_,
+psyplot-gui_, psy-simple_, psy-maps_ and psy-reg_ plus the conda package for
+managing virtual environments. These installers have been created using using
+the conda constructor_ package and the packages from the conda-forge_ channel.
 
 .. only:: html and not epub
 
@@ -49,38 +61,8 @@ psy-maps_ and psy-reg_) and their dependencies.
 Files for all versions can be found in the psyplot-conda_ repository and
 explicitly on the releases_ page.
 
-Installation on MacOSX and Linux
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-Download the bash script (file ending on ``'.sh'`` for linux or MacOSX) from
-above or the releases_ page and open a terminal window.
-
-Type::
-
-    bash '<path-to-the-downloaded-file.sh>'
-
-and simply follow the instructions.
-
-For more informations on the command line options type
-
-Type::
-
-    bash '<Path-to-the-downloaded-file.sh>' --help
-
-On MaxOSX, the installation will create the ``Psyplot.app`` in the
-``/Applications`` folder (or alternatively the``~/Applications`` folder) which
-you can use to open the psyplot GUI. Otherwise type ``psyplot`` in the
-terminal.
-
-
-Installation on Windows
-~~~~~~~~~~~~~~~~~~~~~~~
-Download the executable script (file ending on ``'.exe'``) for your platform
-from above or the releases_ page.
-
-Just double click the downloaded file and follow the instructions. The
[…]
