### 2017-08-12 e3389fb
Improved installation instructions
