### 2017-08-13 #344 @Fil
Yes 👌🏼

PS: Maybe this issue could show the actual URL for the paper <http://joss.theoj.org/papers/3535c28017003f0b5fb63b1b64118b60> ?
