### 2017-08-13 #344 @arfon
> PS: Maybe this issue could show the actual URL for the paper http://joss.theoj.org/papers/3535c28017003f0b5fb63b1b64118b60

That's a good idea. The main review issue (which we're about to switch to) _does_ have the URL linked to from the status badge but I agree it would be good to have that here too.
