### 2017-08-13 #344 @arfon
@Fil - please head over to #363 for the actual review.
