### 2017-08-13 #363 @Fil
@arfon this is my first review, so maybe my question is obvious: where is the "thing" that I'm reviewing? (not the software, which I found, but the paper that will be published — probably something named `paper.md`)
