### 2017-08-13 ffb0a2b
Added defaultParams_orig [skip ci]
