### 2017-08-13 ca325be
Added notes about configuration [skip ci]
