### 2017-08-13 #363 @arfon
Hi @Fil. JOSS reviews are primarily about the software (https://github.com/Chilipp/psyplot.git). We also ask that you check that the associated paper (https://github.com/Chilipp/psyplot/blob/master/paper.md) meets the [recommended structure/content requirements](http://joss.theoj.org/about#paper_structure).
