### 2017-08-13 21e6e6d
Minor changes [skip ci]
