### 2017-08-13 #363 @Fil
Yes absolutely. I'm mostly offline atm so was doing the simple admin things — will do the install later next week
