### 2017-08-13 bdd9d17
Minor changes [skip ci]
