### 2017-08-17 #363 @Fil
@chilipp I've followed the installation procedure (conda in a virtualenv) and I can't seem to get the plot as a graphics. Here's what I get in a jupyter notebook (after `conda install nb_conda`).

I was hoping to get the map as in the example http://psyplot.readthedocs.io/en/latest/getting_started.html

Alternatively, I couldn't find how to launch the psyplot gui (from http://psyplot.readthedocs.io/projects/psyplot-gui/en/latest/getting_started.html).
