### 2017-08-17 #363 @Fil
With psyplot-gui running, I was able to get the map plotted, and change its projection as in the example.

Minor corrections to the install procedure to explain if it is possible to use from jupyter notebook, and how to launch the gui would be welcome.
