### 2017-08-17 34b8f29
Add a note for how to view the plot
Solves
https://github.com/openjournals/joss-reviews/issues/363#issuecomment-322990894
