### 2017-08-17 845f8de
Add a note on how to run getting_started
This change should help to avoid situations as in
https://github.com/openjournals/joss-reviews/issues/363#issuecomment-322989176
[skip ci]
