### 2017-08-17 f70e69d
Minor changes [skip ci]
