### 2017-08-17 #363 @Chilipp
@Fil , thanks a lot for the feedback and your typo corrections!

I added notes to the docs of psyplot, which should solve this problem (see the 2 first notes in the [Getting started](http://psyplot.readthedocs.io/en/latest/getting_started.html) example. Additionally I included a [section in the psyplot-gui docs](http://psyplot.readthedocs.io/projects/psyplot-gui/en/latest/getting_started.html#starting-the-gui) on how to start the GUI.
