### 2017-08-17 #363 @Fil
About the "Community Guidelines" item: the package's README only hints a github issues; it could say a bit more clearly how to contribute, and for example if pull requests, examples, and documentation are welcome.
