### 2017-08-17 #363 @Fil
About "References", please fix the references DOI:
- xarray according to http://xarray.pydata.org/en/stable/faq.html#how-should-i-cite-xarray
- matplotlib according to http://ieeexplore.ieee.org/document/4160265/
