### 2017-08-17 b9951a2
Fixed references
Solves
https://github.com/openjournals/joss-reviews/issues/363#issuecomment-323025143
[skip ci]
