+++ paper.bib
@@ -1,13 +1,15 @@
 @article{Hunter2007,
-  author =    {Hunter, J. D.},
-  title =     {Matplotlib: A 2D graphics environment},
-  journal =   {Computing In Science \& Engineering},
-  year =      {2007},
-  volume =    {9},
-  number =    {3},
-  pages =     {90--95},
-  abstract =  {Matplotlib is a 2D graphics package used for Python for application development, interactive scripting, and publication-quality image generation across user interfaces and operating systems.},
-  publisher = {IEEE COMPUTER SOC}
+    author={J. D. Hunter},
+    journal={Computing in Science Engineering},
+    title={Matplotlib: A 2D Graphics Environment},
+    year={2007},
+    volume={9},
+    number={3},
+    pages={90--95},
+    keywords={computer graphics;mathematics computing;object-oriented programming;software packages;2D graphics package;Matplotlib;Python;application development;interactive scripting;operating system;publication-quality image generation;user interface;Computer languages;Equations;Graphical user interfaces;Graphics;Image generation;Interpolation;Operating systems;Packaging;Programming profession;User interfaces;Python;application development;scientific programming;scripting languages},
+    doi={10.1109/MCSE.2007.55},
+    ISSN={1521-9615},
+    month={May}
 }
 
 @online{psyplot,
@@ -50,11 +52,14 @@
 	urldate = {2017-07-28},
 }
 
-@online{xarray,
-	author = {Hoyer, Stephan and Kleeman, Alex and Brevdo, Eugene},
-	title = {xarray: N-D labeled arrays and datasets in Python},
-	year = 2014,
-	url = {http://xarray.pydata.org/en/stable/},
-	urldate = {2017-07-28},
+@article{hoyer2017xarray,
+  title     = {xarray: {N-D} labeled arrays and datasets in {Python}},
+  author    = {Hoyer, S. and J. Hamman},
+  journal   = {Journal of Open Research Software},
+  volume    = {5},
+  number    = {1},
+  year      = {2017},
+  publisher = {Ubiquity Press},
+  doi       = {10.5334/jors.148},
+  url       = {http://doi.org/10.5334/jors.148}
 }
-
