### 2017-08-17 9d73b13
Added links to github repos for subprojects
