+++ docs/projects.rst
@@ -8,11 +8,34 @@ splitted into several subprojects. Each of them is accessible via
 ``https://psyplot.readthedocs.io/projects/<project-name>``
 
 - the :ref:`psyplot_gui <psyplot_gui:psyplot-gui>` package: The GUI to psyplot
+  |psyplot-gui-github|
 - the :ref:`psy-simple <psy_simple:psy-simple>` package: A plugin for simple
-  visualization
+  visualization |psy-simple-github|
 - the :ref:`psy-maps <psy_maps:psy-maps>` package: A psyplot plugin for
-  visualizing data on a map
+  visualizing data on a map |psy-maps-github|
 - the :ref:`psy-reg <psy_reg:psy-reg>` package: A psyplot plugin for visualizing
-  and calculating regression fits
+  and calculating regression fits |psy-reg-github|
+- the psyplot-conda repository. A repository for the creation of standalone
+  installers for psyplot (see :ref:`install-standalone`) |psyplot-conda-github|
 
 See :ref:`plugins` for more informations on the plugins.
+
+.. |psyplot-gui-github| image:: GitHub-Mark-16px.png
+    :target: https://github.com/Chilipp/psyplot-gui
+    :alt: github
+
+.. |psy-simple-github| image:: GitHub-Mark-16px.png
+    :target: https://github.com/Chilipp/psy-simple
+    :alt: github
+
+.. |psy-maps-github| image:: GitHub-Mark-16px.png
+    :target: https://github.com/Chilipp/psy-maps
+    :alt: github
+
+.. |psy-reg-github| image:: GitHub-Mark-16px.png
+    :target: https://github.com/Chilipp/psy-reg
+    :alt: github
+
+.. |psyplot-conda-github| image:: GitHub-Mark-16px.png
+    :target: https://github.com/Chilipp/psyplot-conda
+    :alt: github
