### 2017-08-17 3ecfbb7
Minor changes [skip ci]
