+++ docs/develop/index.rst
@@ -1,7 +1,7 @@
 .. _developers-guide:
 
 Developers guide
-----------------
+================
 In this section we provide a deeper overview and introduction in the psyplot
 frameworks that is necessary for creating new plugins.
 
