### 2017-08-17 8bd13be
Added sphinx_nbexamples to insterphinx [skip ci]
