+++ docs/conf.py
@@ -234,6 +234,8 @@ intersphinx_mapping = {
     'xarray': ('http://xarray.pydata.org/en/stable/', None),
     'cartopy': ('http://scitools.org.uk/cartopy/docs/latest/', None),
     'mpl_toolkits': ('http://matplotlib.org/basemap/', None),
+    'sphinx_nbexamples': ('http://sphinx-nbexamples.readthedocs.io/en/latest/',
+                          None),
     'psy_maps': (
         'https://psyplot.readthedocs.io/projects/psy-maps/en/latest/', None),
     'psy_simple': (
