### 2017-08-17 f313b3e
Added contributions guidelines
Solves
https://github.com/openjournals/joss-reviews/issues/363#issuecomment-323022927
[skip ci]
