+++ docs/contribute.rst
@@ -1,0 +1,25 @@
+.. _how-to-contribute:
+
+How to contribute
+=================
+psyplot and it's plugins are available open-source on Github. Here we very much
+welcome your contributions!
+
+In case of any troubles, need for clarification or suggestions, please open an
+issue on `psyplots github page`_ (or on the corresponding plugins github page,
+see :ref:`projects`). If you are not sure, where you should open the issue,
+just use psyplots repository.
+
+We also very much welcome pull requests for bug fixes, feature improvements or
+additional examples. If you want to add an example to the documentation, please
+just fork the correct github repository and add a jupyter notebook in the
+``examples`` directory, together with all the necessary data files. To build
+our docs, we use the sphinx-nbexamples_ package to build the examples.
+Therefore please make sure that your supplementary data files are correctly
+implemented in the meta data of the notebook (see
+:ref:`sphinx_nbexamples:supp`). We are, however, also willing to help you
+finalizing incomplete pull requests.
+
+
+.. _psyplots github page: https://github.com/Chilipp/psyplot/issues
+.. _sphinx-nbexamples: http://sphinx-nbexamples.readthedocs.io/en/latest/
