### 2017-08-17 31a5d87
Minor changes [skip ci]
