### 2017-08-18 #363 @Chilipp
@arfon, I will use Zenodo for it. However I encountered a bug on their page. A couple of months ago, the psyplot repository has been deleted and created new from scratch. Now I cannot enable it again for Zenodo.

I contacted Zenodo and would like to wait for their response. Otherwise I could only provide a DOI of the [psyplot-conda](https://github.com/Chilipp/psyplot-conda) repository, which provides the installers for psyplot.
