### 2017-08-19 #363 @arfon
> I contacted Zenodo and would like to wait for their response. Otherwise I could only provide a DOI of the psyplot-conda repository, which provides the installers for psyplot.

Agreed. Let's wait for their response here.
