### 2017-08-21 2070f98
Added acknowledgment for SNF [skip ci]
