### 2017-08-21 #363 @Chilipp
@arfon, I thought a bit more about it and I now think that it makes the most sense to use the DOI for psyplot-conda I just created

[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.845455.svg)](https://doi.org/10.5281/zenodo.845455)

This would be the best for the JOSS paper, I think, because this archive then also includes the plugins psy-simple, psy-maps and psy-reg and the GUI, plus the installers, which have also been tested by @Fil.
