### 2017-08-22 #363 @arfon
@Fil - many thanks for your review here ✨

@Chilipp - your paper now accepted into JOSS and your DOI is http://dx.doi.org/10.21105/joss.00363 ⚡️ 🚀 💥
